Clean up + bug fix

+++ main.py
@@ -6,6 +6,7 @@ import random
 
 
 token= ''  
+
 command_prefix_global = '--- '
 
 
@@ -63,21 +64,22 @@ async def on_message(message):
         attachment = message.attachments[0]
         content = await attachment.read()  # Odczytuje zawartość jako bajty
         text_content = content.decode('utf-8')  
-        await process_file_content(message,text_content) ##Dodac async w funckji
-        await message.delete()
-        return
+        if '1.' in text_content and '2.' in text_content:
+            await process_file_content(message,text_content) ##Dodac async w funckji
+            await message.delete()
+            return
+        else:
+            return
 
     elif isinstance(message.channel, discord.TextChannel) and message.content.startswith(command_prefix_global):
         await create_thread(message, message.content.split(' ')[1])
         await message.delete()
         return
     
-
     elif isinstance(message.channel, discord.Thread) and (message.content.startswith('+') or message.content.startswith('-')):
         await change_thread_description(message, message.content)
         await message.delete()
         return
-    
 
     
     return 
@@ -91,6 +93,7 @@ async def on_message(message):
 
 async def create_thread(ctx, thread_name: str):
     existing_thread = discord.utils.find(lambda t: t.name == thread_name, ctx.channel.threads)
+
     description = description_template.format(
         first_row='',
         second_row='',
@@ -107,7 +110,6 @@ async def create_thread(ctx, thread_name: str):
     initial_message = await ctx.channel.send(description or "No description provided.")
     thread = await initial_message.create_thread(name=thread_name, auto_archive_duration=thread_lifetime) 
     await thread.add_user(ctx.author)
-
 
 async def change_thread_description(ctx, new_comment: str):
     if isinstance(ctx.channel, discord.Thread) and check_pattern(new_comment):
@@ -138,15 +140,7 @@ async def change_thread_description(ctx, new_comment: str):
                                 rows[i][j] = nick
                             else:
                                 rows[i].append(nick)
-
-
-
-                ##
-
                 desc_list = row_modyfication(rows,new_comment,ctx.author.mention) ##
-                #desc = row_modyfication(rows,new_comment,ctx.author.display_name) 
-
-
                 thread_desc = file_description_template.format(
                     thread_author = thread_author_name,
                     description = ''.join(thread_description),
@@ -162,17 +156,9 @@ async def change_thread_description(ctx, new_comment: str):
                     tenth_row=','.join(desc_list[9]) if desc_list[9][0] != '' and desc_list[9][0] != None else '',
                 )
                 await parent_message.edit(content=thread_desc)
-
-
-
-
-
             else:
                 desc = modify_description(parent_message.content,new_comment,ctx.author.mention)
                 await parent_message.edit(content=desc)
-
-
-
 async def process_file_content(ctx,content):
     text_rows = [row for row in content.split('\n') if row != '']
     thread_author_name = ''  ## Pamietac ze to moze zostać ''
@@ -194,8 +180,6 @@ async def process_file_content(ctx,content):
                 pass
         elif table_start_index == 0 and 'Comp:' in single_row:
             table_start_index = index+1
-            #text_rows.index(single_row) + 1
-
 
     if createdby_included:
         thread_description = text_rows[1:table_start_index-1]
@@ -206,18 +190,13 @@ async def process_file_content(ctx,content):
             thread_author_name= nickName.mention
             thread_author = nickName
         else:
-            thread_author_name = '<@nox2210>'                       ### Test
-            thread_author = ctx.guild.get_member('2210')  ## Tp jest zle
-
+            thread_author_name = '<@969918981660086342>'
 
     else:
         thread_description = thread_description.join(text_rows[0:table_start_index-1])  
-        thread_author_name = '<@nox2210>' 
-        thread_author = ctx.guild.get_member('2210')  ## Tp jest zle
+        thread_author_name = '<@969918981660086342>' 
 
-
-    ## Sprawdzenie czy thread  ame jest dodany
-    if '10.' in text_rows[-1]: # Brakuje nam Thread Name w ostatnim watku
+    if '10.' in text_rows[-1]: 
         thread_name= 'Thread_' + str(random.randint(0, 1000))
         for single_row in text_rows[table_start_index:]:
             nicks_list_prep = single_row.split('.')[1].strip().split(',')
@@ -229,7 +208,6 @@ async def process_file_content(ctx,content):
             rows_nick_list.append(nicks_list_prep)
 
     discord_user_list = get_discord_user_list(ctx.guild,rows_nick_list)
-    
 
     thread_desc = file_description_template.format(
         thread_author = thread_author_name,
@@ -245,10 +223,8 @@ async def process_file_content(ctx,content):
         ninth_row=','.join(discord_user_list[8]) if discord_user_list[8][0] !='' and discord_user_list[8][0] != None else '',
         tenth_row=','.join(discord_user_list[9]) if discord_user_list[9][0] != '' and discord_user_list[9][0] != None else '',
     )
-
     initial_message = await ctx.channel.send(thread_desc or "No description provided.")
     thread = await initial_message.create_thread(name=thread_name, auto_archive_duration=thread_lifetime) 
-    #sawait thread.add_user(ctx.author)
 def get_discord_user_list(guild,user_list):
     discord_user_list = []
     for row in user_list:
@@ -266,7 +242,6 @@ def get_discord_user_list(guild,user_list):
                 pass
         discord_user_list.append(discord_user_row)
     return discord_user_list
-
 def modify_description(description,given_comment = '+1.1', given_nick = 'Test'):
     rows = [[None], [None], [None], [None], [None], [None], [None], [None], [None], [None]]
     if description[-1] == ':': description = description+ ' ' #To jeszcze do sprawdzenia
@@ -294,17 +269,6 @@ def modify_description(description,given_comment = '+1.1', given_nick = 'Test'):
         ninth_row=','.join(rows[8]) if rows[8][0] !='' and rows[8][0] != None else '',
         tenth_row=','.join(rows[9]) if rows[9][0] != '' and rows[9][0] != None else '',
     )
-
-
-
-
-
-
-
-
-
-
-
 def row_modyfication(table,comment,nick):
     action_type = comment[0]
     affected_row = int(comment[1:].split('.')[0]) -1  
@@ -332,68 +296,6 @@ def row_modyfication(table,comment,nick):
             table[affected_row][0] =''
 
     return table
-# def modify_description(description,given_comment = '+1.1', given_nick = 'Test'):
-#     rows = [[None], [None], [None], [None], [None], [None], [None], [None], [None], [None]]
-#     if description[-1] == ':': description = description+ ' ' #To jeszcze do sprawdzenia
-#     description_rows = description.split("\n") 
-#     for i in range(10):
-#         extracted_nicks = description_rows[i].split(": ")[1].split(",")
-#         if len(extracted_nicks) > 0:
-#             for j, nick in enumerate(extracted_nicks):
-#                 if len(rows[i]) > j:
-#                     rows[i][j] = nick
-#                 else:
-#                      rows[i].append(nick)
-                
-#     rows = row_modyfication(rows,given_comment,given_nick)
-
-#     return description_template.format(
-#         first_row= ','.join(rows[0]) if rows[0][0] != '' and rows[0][0] != None else '',
-#         second_row= ','.join(rows[1]) if rows[1][0] != '' and rows[1][0] != None else '',
-#         third_row= ','.join(rows[2]) if rows[2][0] != '' and rows[2][0] != None else '',
-#         fourth_row=','.join(rows[3]) if rows[3][0] != '' and rows[3][0] != None else '',
-#         fifth_row=','.join(rows[4]) if rows[4][0] != '' and rows[4][0] != None else '',
-#         sixth_row=','.join(rows[5]) if rows[5][0] != '' and rows[5][0] != None else '',
-#         seventh_row=','.join(rows[6]) if rows[6][0] != '' and rows[6][0] != None else '',
-#         eighth_row=','.join(rows[7]) if rows[7][0] != '' and rows[7][0] != None else '',
-#         ninth_row=','.join(rows[8]) if rows[8][0] !='' and rows[8][0] != None else '',
-#         tenth_row=','.join(rows[9]) if rows[9][0] != '' and rows[9][0] != None else '',
-#     )
-# def row_modyfication(table,comment,nick):
-#     action_type = comment[0]
-#     affected_row = int(comment[1:].split('.')[0]) -1  
-#     affected_positiom = int(comment[1:].split('.')[1]) -1
-    
-#     if action_type == '+':
-#         if table[affected_row] == ['']:
-#             table[affected_row][0] = nick 
-#         else:
-#             if nick not in table[affected_row]:
-#                 if affected_positiom < len(table[affected_row]):
-#                     table[affected_row].insert(affected_positiom,nick)
-#                 else:
-#                     table[affected_row].append(nick)
-#             else:
-#                 table[affected_row].remove(nick)
-#                 table[affected_row].insert(affected_positiom,nick)
-#         return table
-    
-#     elif action_type == '-':
-#         table_affected_row_len = len(table[affected_row])
-#         if table[affected_row] != [''] and table_affected_row_len != 1 and table_affected_row_len > affected_positiom :
-#             table[affected_row].pop(affected_positiom) 
-#         elif table_affected_row_len == 1 and affected_positiom == 0: 
-#             table[affected_row][0] =''
-
-#     return table
-
-
-
-
-
-
-
-
 def check_pattern(s):
     if s and s[0] in ['+', '-']:
         pattern = r'^[+-]?\d+\.\d+$'
